+++ src/components/Editors/Metadata/Layout.tsx
@@ -3,13 +3,22 @@ import Box from '@mui/material/Box'
 import Actions from './Actions'
 import Content from './Content'
 import Header from './Header'
+import Welcome from './Welcome'
+import { useStore } from './store'
 
 export default function Layout() {
+  const editor = useStore((state) => state.editorState.editor)
   return (
     <Box>
       <Header />
-      <Content />
-      <Actions />
+      {editor ? (
+        <Box>
+          <Content />
+          <Actions />
+        </Box>
+      ) : (
+        <Welcome />
+      )}
     </Box>
   )
 }
